+++ CHANGELOG.md
@@ -14,6 +14,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.0.0] - 2023-12-31
 ### Added
+
+- `folder` objective
 - Logging
     - Ingame logging
         - Logging during `/q reload`